fix(events-processor): filter out incoming nulls
incoming dynamodb stream events are tranformed to a common message format, which does not support
nested json at this time, therefore nested json attributes were being treated as null. when this was
added to the rules engine as a fact, an error was thrown as a fact must have a value. now all
transformed attributes that have an undefined / null value are removed before updating the rules
engine.