fix(assetlibrary): fixed pagination for related items
list of related items is built from a few different graph traversals
which are then projects into a single result. pagination was applied to
each traversal before projecting.  it was possible the graph traversals
would return different results to each other due to different filtering
being applied.  This has been refactored to simplify, and synchornize
the filtering of the individual traversals.