fix(cdf): reduced lambda bundle size
reconfiguring the dependencies of all cdf libraries to peer dependencies allows for shared reuse of
dependencies at the top level thus reducing the lambda bundle size.  What this means though is that
for each consuming service of a cdf library, such as a CDF core service or a facade, the
dependencies that once existed at the library level now need explicitly adding as a dependency of
the consuming service.