fix(events): dAX subnet group
A recent change introduced a backwards incompatible change. Introduction of an explicit subnet for
the DAX cluster would require the DAX resource being deleted and recreated, but this is referenced
by the events-alerts stack meaning the resource cannot be deleted without the events-alerts stack
being deleted first. The original breaking change has been modified so that the use of a custom
subnet is optional, thus not affecting existing deployments.